IToStr and DisplayString, Burn Rows and Game Fail

+++ base_types.h
@@ -11,7 +11,7 @@ Description: <empty>
 
 #include <stdint.h>
 
-typedef uint32_t u64;
+typedef uint64_t u64;
 typedef uint32_t u32;
 typedef uint16_t u16;
 typedef uint8_t u8;
